refactor: update file names and folder structure
