+++ src/interfaces/animal.ts
