+++ src/router/index.tsx
@@ -1,0 +1,3 @@
+import router from './router';
+
+export default router;
